Move symbol example to its own new example class. Add centered pin.
