+++ Apps/Examples/Examples/Models/Examples.swift
@@ -39,6 +39,9 @@ public struct Examples {
         Example(title: "Select an annotation",
                 description: "Select an annotation with a tap gesture.",
                 type: SelectAnnotationExample.self),
+        Example(title: "Add anchored annotations with a custom map layer.",
+                description: "Add anchored image annotations to the map with a custom SymbolLayer. Annotations are styled via feature attributes.",
+                type: CustomSymbolAnnotationsExample.self),
         Example(title: "Add a GeoJSON data source",
                 description: "Render multiple geometries from GeoJSON data on the map.",
                 type: GeoJSONSourceExample.self),
